Added BookList.razor
In 'BookList.razor', enabled the display of the username on the booklist page. This was achieved by injecting the UserManager and AuthenticationStateProvider and creating a User property to store the currently logged in user's information. If the username is null, it fetches the current user details. If the username is provided, it looks up the user with that username. If no such user exists, it sets the Username property to "Not Found".