Add Bootstrap for styling and a modal for book editing and made Books editable in BookList.razor
Bootstrap package reference was added to the BookTracker App csproj file to improve the UI styling. A link to Bootstrap CDN and jQuery were also added to the _Host.cshtml page. A tabular form was created in the BookList.razor page for the book list and an edit button was also added by utilising a Bootstrap modal, allowing the user to edit their book details directly from the page. This will offer a better user experience and make editing book details quick and convenient for users. Finally, a method was added to save the book details after editing.