Add create book modal and fix input id
Added a modal for creating new books, which significantly improves the usability of the application. Added additional input fields in the modal for the 'title', 'author', 'pages', 'pages read', and 'rating' fields for the Book object. The addition of these fields helps improve the user experience by providing more detailed input options.

Additionally, an error in the id attribute of the 'pages' input field on the book edit form was corrected. This fix is critical as it ensures the correct binding of the pages input field to the Book object. Now the form works as expected.