Add EditBookModal and refactor existing code for cleanliness
A new component EditBookModal has been added which allows users to edit a book's details in a dedicated modal, improving both usability and code organisation. Refactoring was also done on BookList.razor and AddBookModal.razor for better code readability and separation of concerns. The IBookManagementService has been updated with a DeleteBook method for future functionalities.