Refactor ApplicationUser deletion logic in UserManager
In the ApplicationUserManager.cs file, the user deletion logic has been refactored. Initially, the user was removed from the database without any preliminary checks or user feedback. The DeleteUser method has been updated so that it now returns a result of type IdentityResult to provide details on whether the operation has succeeded or failed. This change allows for a more robust handling of errors that may occur during the execution of this operation. The method interfaces have also been updated across the application to reflect this change. The DeletePersonalData.cshtml.cs now utilises the updated DeleteUser method from ApplicationUserManager.