Remove redundant files and UI refactoring
Removed several unnecessary files (Counter, FetchData, NavMenu, SurveyPrompt) and refactored the Index and nav bar. The removal was necessary to minimize clutter and improve navigation. NavMenu moved to Navbar and Index was refactored to make it more informative and user-oriented. The UI redesign was aimed to enhance user experience and better output display. New CSS file for Navbar was added for handling design-related changes.