Remove unnecessary bin and obj folders
Deleted bin and obj folders which are not required in the repository. These folders contain build output and temporary files created by the compiler. Therefore, these files are typically not included in the version control system. Including them bloats the repository size and can cause conflicts with other developer's work.